Bot added and some other edits - TgBot some arch added - Animal struct created and table added to database - DB_Password added to .env - User fields edited to take lower memory

+++ internal/users/user.go
@@ -1,9 +1,9 @@
 package users
 
 type User struct {
-	Id         int64  `json:"Id"`
+	Id         int    `json:"Id"`
 	TelegramId string `json:"TelegramId"`
 	Login      string `json:"Login"`
 	Password   string `json:"Password"`
-	AnimalId   int64  `json:"AnimalId"`
+	AnimalId   int    `json:"AnimalId"`
 }
